Implemented screen scrolling.

+++ common.h
@@ -41,7 +41,7 @@ typedef struct registers
 
 
 // functions
-void memcpy(uint8_t *dest, uint8_t *src, uint32_t len);
+void memcpy(uint16_t *dest, uint16_t *src, uint32_t len);
 void memset(uint8_t *dest, uint8_t val, uint32_t len);
 
 #endif
